padronização do retorno de permission

+++ CoreLogin_Presentation/Controllers/PermissionController.cs
@@ -1,8 +1,10 @@
-﻿using CoreLogin_Domain.Converters;
+﻿using CoreLogin.Domain.Entities.Enum;
+using CoreLogin_Domain.Converters;
 using CoreLogin_Domain.Converters.DTO;
 using CoreLogin_Domain.Entities;
 using CoreLogin_Domain.Repositories;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.Annotations;
@@ -24,9 +26,20 @@ namespace CoreLogin_Presentation.Controllers
     [AllowAnonymous]
     public async Task<ActionResult<PermissionResultDTO>> AddPermissionAsync([FromBody] PermissionRequestDTO permission)
     {
+      // validate if the permission.type is valid
+      if (!Enum.IsDefined(typeof(EPermissionType), permission.Type))
+      {
+        return new BadRequestObjectResult("Invalid Permission Type");
+      }
+
       var newPermission = await _permissionRepository.CreatePermissionAsync(permission);
 
-      return newPermission;
+      if(newPermission == null)
+      {
+        return new BadRequestObjectResult("Error on create new permission");
+      }
+
+      return new OkObjectResult(newPermission);
     }
 
     [HttpGet]
@@ -44,7 +57,20 @@ namespace CoreLogin_Presentation.Controllers
     [AllowAnonymous]
     public async Task<ActionResult> DeletePermissionAsync([FromBody] PermissionRequestDTO permission)
     {
-      return await _permissionRepository.DeletePermissionAsync(permission);
+
+      if (!Enum.IsDefined(typeof(EPermissionOperation), permission.Operation) || !Enum.IsDefined(typeof(EPermissionType), permission.Type))
+      {
+        return new BadRequestObjectResult("Invalid 'Permission Operation' or 'Type'");
+      }
+
+      var permissionDeleted = await _permissionRepository.DeletePermissionAsync(permission);
+
+      if (permissionDeleted == null)
+      {
+        return new BadRequestObjectResult("Permission not found");
+      }
+
+      return new OkResult();
     }
 
   }
